--- src/main/java/duke/command/Command.java
@@ -1,28 +1,0 @@
-package duke.command;
-
-import duke.DukeException;
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command to be executed.
- */
-public abstract class Command {
-
-    /**
-     * Executes the command.
-     * @param tasks The list of tasks.
-     * @param ui The user interface to interact with the user.
-     * @param storage The storage to save the tasks.
-     * @throws DukeException If an error occurs during the execution of the command.
-     */
-    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
-
-    /**
-     * Returns whether the command is an exit command.
-     * @return Whether the command is an exit command.
-     */
-    public abstract boolean isExit();
-}
-
