--- src/main/java/duke/command/DeleteCommand.java
@@ -1,41 +1,0 @@
-package duke.command;
-
-import duke.DukeException;
-import duke.storage.Storage;
-import duke.task.Task;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command to delete a task.
- */
-public class DeleteCommand extends Command {
-    protected int index;
-
-    /**
-     * Constructor for DeleteCommand.
-     * @param index The index of the task to be deleted.
-     */
-    public DeleteCommand(int index) {
-        this.index = index;
-    }
-
-    /**
-     * Executes the command to delete a task.
-     * @param tasks The list of tasks.
-     * @param ui The user interface to interact with the user.
-     * @param storage The storage to save the tasks.
-     * @throws DukeException If an error occurs during the execution of the command.
-     */
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        Task task = tasks.getTask(index);
-        tasks.deleteTask(index);
-        ui.showTaskDeleted(task, tasks.getSize());
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
