--- src/main/java/duke/command/TodoCommand.java
@@ -1,39 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.Todo;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command to add a todo task.
- */
-public class TodoCommand extends Command {
-    protected String description;
-
-    /**
-     * Constructor for TodoCommand.
-     * @param description The description of the todo.
-     */
-    public TodoCommand(String description) {
-        this.description = description;
-    }
-
-    /**
-     * Executes the command to add a todo task.
-     * @param tasks The list of tasks.
-     * @param ui The user interface to interact with the user.
-     * @param storage The storage to save the tasks.
-     */
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        Todo todo = new Todo(description);
-        tasks.addTask(todo);
-        ui.showTaskAdded(todo, tasks.getSize());
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
