--- src/main/java/duke/command/UnmarkCommand.java
@@ -1,39 +1,0 @@
-package duke.command;
-
-import duke.DukeException;
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-/**
- * Represents a command to unmark a task as done.
- */
-public class UnmarkCommand extends Command {
-    private int index;
-
-    /**
-     * Constructor for UnmarkCommand.
-     * @param index The index of the task to be marked as not done.
-     */
-    public UnmarkCommand(int index) {
-        this.index = index;
-    }
-
-    /**
-     * Executes the command to mark a task as not done.
-     * @param tasks The list of tasks.
-     * @param ui The user interface to interact with the user.
-     * @param storage The storage to save the tasks.
-     * @throws DukeException If an error occurs during the execution of the command.
-     */
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        tasks.getTask(index).markAsNotDone();
-        ui.showMessage("Nice! I've marked this task as undone:\n" + tasks.getTask(index));
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
