+++ src/main/java/fluffy/Fluffy.java
@@ -1,0 +1,19 @@
+package fluffy;
+
+import javafx.application.Application;
+
+/**
+ * The Fluffy program is a chatbot that helps users to keep track of their tasks.
+ * It can store and retrieve tasks from a file, and perform operations such as adding, deleting,
+ * and marking tasks as done.
+ */
+public class Fluffy {
+
+    /**
+     * Main method to run the Fluffy program.
+     * @param args The command line arguments.
+     */
+    public static void main(String[] args) {
+        Application.launch(FluffyApp.class);
+    }
+}
