--- src/test/java/duke/DukeTest.java
@@ -1,2 +1,0 @@
-package duke;
-
