+++ src/test/java/fluffy/FluffyTest.java
@@ -1,0 +1,2 @@
+package fluffy;
+
